add Home, Camera and Messages fragments to Home Package
