+++ app/src/main/java/jocontreras77/com/fiskagramapp/Home/CameraFragment.java
@@ -1,0 +1,26 @@
+package jocontreras77.com.fiskagramapp.Home;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import jocontreras77.com.fiskagramapp.R;
+
+
+/**
+ * Created by Jennifer on 8/29/17.
+ */
+
+public class CameraFragment extends Fragment{
+    private  static final String TAG = "CameraFragment";
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_camera, container, false);
+        return view;
+    }
+}
