+++ app/src/main/java/jocontreras77/com/fiskagramapp/Home/HomeFragment.java
@@ -1,0 +1,27 @@
+package jocontreras77.com.fiskagramapp.Home;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import jocontreras77.com.fiskagramapp.R;
+
+
+
+/**
+ * Created by Jennifer on 8/29/17.
+ */
+
+public class HomeFragment extends Fragment{
+    private  static final String TAG = "HomeFragment";
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_home, container, false);
+        return view;
+    }
+}
