+++ app/src/main/java/jocontreras77/com/fiskagramapp/Profile/ProfileActivity.java
@@ -1,4 +1,4 @@
-package jocontreras77.com.fiskagramapp;
+package jocontreras77.com.fiskagramapp.Profile;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -10,16 +10,17 @@ import android.view.MenuItem;
 
 import com.ittianyu.bottomnavigationviewex.BottomNavigationViewEx;
 
+import jocontreras77.com.fiskagramapp.R;
 import jocontreras77.com.fiskagramapp.Utils.BottomNavigationViewHelper;
 
 /**
  * Created by Jennifer on 8/29/17.
  */
 
-public class LikesActivity extends AppCompatActivity {
-    private static final String TAG = "LikesActivity";
-    private static final int ACTIVITY_NUM = 3;
-    private Context mContext = LikesActivity.this;
+public class ProfileActivity extends AppCompatActivity {
+    private static final String TAG = "ProfileActivity";
+    private static final int ACTIVITY_NUM = 4;
+    private Context mContext = ProfileActivity.this;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
