+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="jocontreras77.com.fiskagramapp.HomeActivity">
+    tools:context="jocontreras77.com.fiskagramapp.Home.HomeActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
